refactor: split oversized views into one-view-per-file units
