+++ dogArea/Views/GlobalViews/PositionMarkerViewWithSelection.swift
@@ -1,17 +1,12 @@
-//
-//  PositionMarkerView.swift
-//  dogArea
-//
-//  Created by 김태훈 on 11/8/23.
-//
-
 import SwiftUI
 
-struct PositionMarkerView: View {
+struct PositionMarkerViewWithSelection: View {
+    @State var selected: Bool
+
     var body: some View {
         ZStack {
             RoundedRectangle(cornerRadius: 5)
-                .fill(Color.appYellowPale)
+                .fill(selected ? Color.appPeach : Color.appYellowPale)
             Text("💦").font(.appFont(for: .Bold, size: 10))
                 .padding(5)
         }
