+++ dogArea/Views/HomeView/HomeSubView/HomeGoalMetricColumnView.swift
@@ -1,0 +1,27 @@
+import SwiftUI
+
+struct HomeGoalMetricColumnView: View {
+    let title: String
+    let value: String
+    let detail: String
+
+    var body: some View {
+        VStack(alignment: .leading, spacing: 6) {
+            Text(title)
+                .font(.appScaledFont(for: .Regular, size: 12, relativeTo: .caption))
+                .foregroundStyle(Color.appDynamicHex(light: 0xC2410C, dark: 0xFDBA74))
+            Text(value)
+                .font(.appScaledFont(for: .SemiBold, size: 34, relativeTo: .title3))
+                .foregroundStyle(Color.appDynamicHex(light: 0x7C2D12, dark: 0xFEF3C7))
+                .lineLimit(1)
+                .minimumScaleFactor(0.7)
+            Text(detail)
+                .font(.appScaledFont(for: .Regular, size: 11, relativeTo: .caption2))
+                .foregroundStyle(Color.appDynamicHex(light: 0x92400E, dark: 0xFDBA74))
+                .lineLimit(2)
+        }
+        .frame(maxWidth: .infinity, alignment: .leading)
+        .accessibilityElement(children: .ignore)
+        .accessibilityLabel("\(title) \(value) \(detail)")
+    }
+}
